added add_module_test
